add batch and threads for indexing mail to matching entity, corrected error generating pdf file with attachment. Remaining corretion todo : batch INDEX_MAIL_TO_ENTITY not executing automatically, associate thread class java not found

+++ src/main/java/com/jss/osiris/libs/mail/GeneratePdfDelegate.java
@@ -1464,7 +1464,7 @@ public class GeneratePdfDelegate {
         return tempFile;
     }
 
-    public File generateGenericFromHtml(String htmlContent, Long mailId)
+    public File generateGenericFromHtml(String htmlContent, Integer mailId)
             throws OsirisException, OsirisValidationException, OsirisClientMessageException {
         File tempFile;
         OutputStream outputStream;
